data-engineering-project-2024/forex-consumer/requirements.txt
kafka-python==2.0.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9